Create README.md
I'm dedicated computer science and engineering student at Unity University, blending theoretical knowledge with practical skills. Proficient in web development, mobile application development, and system development and so much more. Eager to contribute expertise to collaborative projects, I'm on a journey to shape the future of the tech landscape. With a relentless curiosity and a passion for problem-solving, embodies the spirit of a forward-thinking technologist ready to make a meaningful impact.